+++ packages/cldr-core/src/api/general.ts
@@ -1,11 +1,10 @@
 import {
+  AltType,
   CharacterOrderType,
   ContextType,
-  LanguageIdType,
   LineOrderType,
   NamesSchema,
-  RegionIdType,
-  ScriptIdType,
+  Vector2Arrow
 } from '@phensley/cldr-types';
 
 import { MessageFormatter, MessageFormatterOptions } from '@phensley/messageformat';
@@ -29,6 +28,18 @@ import { PrivateApiImpl } from './private/api';
 import { ContextTransformInfo } from '../common/private';
 
 const DEFAULT_NAME_OPTIONS: DisplayNameOptions = { context: 'begin-sentence' };
+
+// Compound fields for language display name
+type LanguageTagFunc = (t: LanguageTag) => string;
+
+const F_LANG_REGION = (t: LanguageTag) => `${t.language()}-${t.region()}`;
+const F_LANG_SCRIPT = (t: LanguageTag) => `${t.language()}-${t.script()}`;
+
+const LANGUAGE_FUNCS: LanguageTagFunc[] = [
+  F_LANG_REGION,
+  F_LANG_SCRIPT,
+  t => t.language()
+];
 
 export class GeneralImpl implements General {
 
@@ -122,20 +133,94 @@ export class GeneralImpl implements General {
     return this.general.formatListToParts(this._bundle, items, type || 'and');
   }
 
-  getLanguageDisplayName(code: string, options: DisplayNameOptions = DEFAULT_NAME_OPTIONS): string {
-    const s = this.names.languages.displayName.get(this._bundle, code as LanguageIdType);
+  getLanguageDisplayName(code: string | LanguageTag, options: DisplayNameOptions = DEFAULT_NAME_OPTIONS): string {
+    const type = options.type || 'none';
+    const arrow = this.names.languages.displayName;
+
+    let tag: LanguageTag = typeof code === 'string' ? parseLanguageTag(code) : code;
+    let s: string = '';
+
+    // First attempt to match the exact string
+    if (typeof code === 'string') {
+      s = this._getVectorAlt(arrow, code, type);
+    }
+    // Try language + region
+    if (!s && tag.hasLanguage() && tag.hasRegion()) {
+      s = this._getVectorAlt(arrow, F_LANG_REGION(tag), type);
+    }
+    // Try language + script
+    if (!s && tag.hasLanguage() && tag.hasScript()) {
+      s = this._getVectorAlt(arrow, F_LANG_SCRIPT(tag), type);
+    }
+    // Try language if script and region are empty
+    if (!s && !tag.hasScript() && !tag.hasRegion()) {
+      s = this._getVectorAlt(arrow, tag.language(), type);
+    }
+
+    // Resolve to fill in unknown subtags, then attempt combinations
+    if (!s) {
+      const locale = this.resolveLocale(tag);
+      tag = locale.tag;
+      for (const func of LANGUAGE_FUNCS) {
+        const id = func(tag);
+        s = this._getVectorAlt(arrow, id, type);
+        if (s) {
+          // Found one
+          break;
+        }
+      }
+    }
     return this.general.contextTransform(s, this.transform, _ctx(options), 'languages');
   }
 
-  getScriptDisplayName(code: string, options: DisplayNameOptions = DEFAULT_NAME_OPTIONS): string {
-    const s = this.names.scripts.displayName.get(this._bundle, code as ScriptIdType);
+  getScriptDisplayName(code: string | LanguageTag, options: DisplayNameOptions = DEFAULT_NAME_OPTIONS): string {
+    const arrow = this.names.scripts.displayName;
+    const type = options.type || 'none';
+    let s: string = '';
+    if (typeof code === 'string') {
+      s = this._getVectorAlt(arrow, code, type);
+
+    // If language is blank or we have an explicit script subtag, use the
+    // script subtag as-is. This will resolve "und-Zzzz" to "Unknown" but
+    // "en-Zzzz" will fall through to resolve "Latin"
+    } else if (!code.hasLanguage() || code.hasScript()) {
+      s = this._getVectorAlt(arrow, code.script(), type);
+    }
+    if (!s) {
+      // Resolve to populate the script
+      const locale = this.resolveLocale(code);
+      s = this._getVectorAlt(arrow, locale.tag.script(), type);
+    }
     return this.general.contextTransform(s, this.transform, _ctx(options), 'script');
   }
 
-  getRegionDisplayName(code: string, options: DisplayNameOptions = DEFAULT_NAME_OPTIONS): string {
-    const impl = this.names.regions.displayName;
-    const name = impl.get(this._bundle, _def(options, 'type', 'none'), code as RegionIdType);
-    return name ? name : impl.get(this._bundle, 'none', code as RegionIdType);
+  getRegionDisplayName(code: string | LanguageTag, options: DisplayNameOptions = DEFAULT_NAME_OPTIONS): string {
+    const arrow = this.names.regions.displayName;
+    const type = options.type || 'none';
+    let s: string = '';
+    if (typeof code === 'string') {
+      s = this._getVectorAlt(arrow, code, type);
+
+    // If language is blank or we have an explicit region subtag, use
+    // the region subtag as-is. This will resolve "und-ZZ" to "Unknown" but
+    // "en-Zzzz" will fall through to resolve "United States"
+    } else if (!code.hasLanguage() || code.hasRegion()) {
+      s = this._getVectorAlt(arrow, code.region(), type);
+    }
+    if (!s) {
+      // Resolve to populate the region
+      const { tag } = this.resolveLocale(code);
+      s = this._getVectorAlt(arrow, tag.region(), type);
+    }
+    // No context transform for region
+    return s;
+  }
+
+  // Check if the given alt type field exists, and fall back to alt type 'nonee'
+  protected _getVectorAlt<T extends string>(arrow: Vector2Arrow<AltType, T>,
+    code: string, type: AltType): string {
+    return arrow.get(this._bundle, type, code as unknown as T)
+      || arrow.get(this._bundle, 'none', code as unknown as T);
   }
 
 }