+++ packages/cldr-schema/src/schema/enums.ts
@@ -3,4 +3,4 @@ import { KeyIndexImpl } from '../instructions';
 
 export const PluralIndex = new KeyIndexImpl<PluralType>(['other', 'zero', 'one', 'two', 'few', 'many']);
 
-export const AltIndex = new KeyIndexImpl(['none', 'short', 'narrow', 'variant', 'stand-alone']);
+export const AltIndex = new KeyIndexImpl(['none', 'short', 'narrow', 'variant', 'stand-alone', 'long', 'menu']);