+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [UNKNOWN]
+### Fixed/Changed
+* API for display names now supports passing either a `string` or `LanguageTag` argument
+* Language display names now uses the script and region subtags for matches.
+* All display name methods will attempt a best match, falling back to matching with likely subtags added.
+
 ## [0.25.2]
 ### Fixed/Changed
 * Ensure formatted timezone offset minutes is an integer